Fixed Seek/Play rengo circular dependency
